app/ocr/dbnet_model.py
import torch
import numpy as np
import cv2
import yaml
import os

def load_dbnet(model_path, cfg_path, device):
    # TODO: Implement actual DBNet model loading
    # You need to import your DBNet model architecture here
    # Example:
    # from your_dbnet_package.model import DBNet
    # 
    # with open(cfg_path, "r") as f:
    #     cfg = yaml.safe_load(f)
    # model = DBNet(cfg['model'])
    # ckpt = torch.load(model_path, map_location=device)
    # if 'state_dict' in ckpt:
    #     model.load_state_dict(ckpt['state_dict'])
    # else:
    #     model.load_state_dict(ckpt)
    # model.to(device)
    # model.eval()
    
    # Placeholder: return None for now
    print(f"⚠️ DBNet model loading is not yet implemented")
    print(f"   Model path: {model_path}")
    print(f"   Config path: {cfg_path}")
    return None

def detect_boxes(dbnet_model, img):
    # TODO: Implement actual detection logic
    # Placeholder: return whole image as single bbox
    h, w = img.shape[:2]
    return [(0, 0, w, h)]